requirements.txt
openai
langchain-core
langchain-openai
PyPDF2
python-dotenv
langchain-community
streamlit
